pyproject.toml
[project]
name = "chat-bot-temp"
version = "0.1.0"
description = "chat bot with temporary memory"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "=3.11.4"
dependencies = [
    "langchain (>=0.2.11,<0.3.0)",
    "langchain-openai (>=0.1.20,<0.2.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "jupyterlab (>=4.2.4,<5.0.0)",
    "langchain-community (>=0.2.10,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
